+++ Docs/Whats_New.md
@@ -27,8 +27,12 @@ Released TBD
 - Fix image subresource sizing calculations for heap-based textures.
 - Fix `MTLHeap` memory leak in `MVKDeviceMemory`.
 - Fix tessellation break when control stage declares but does not use position builtin.
+- Fix inconsistency in reporting device local memory between type and heap on macOS.
+- Fix bug where dynamic shader buffers are overflowing.
+- Avoid reading env vars inside library constructor functions.
+- Update `VK_MVK_MOLTENVK_SPEC_VERSION` to `23`.
 - Cube demo use `VK_EXT_metal_surface` extension.
-- Support *Xcode 11.2*.
+- Support *Xcode 11.3*.